Reverted restructuring of source code; updated where tests are located and how they are identified by pytest
