+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+python_files = *_spec.py
+python_classes = *Spec
