+++ src/app/archbishop/archbishop_spec.py
@@ -13,7 +13,7 @@ def fake_envs(monkeypatch):
 @pytest.fixture()
 def app():
     # required here to avoid premature creation - is create_app being interpreted when imported?
-    from wafflebot import create_app
+    from src.app.app import create_app
 
     return create_app()
