+++ src/app/model/group/Record.py
@@ -1,4 +1,4 @@
-from wafflebot.model.base import Base
+from src.app.model.base import Base
 
 
 class Record(Base):
